feat: move import-agent into create namespace
