+++ src/bedrock_agentcore_starter_toolkit/cli/create/import_agent/agent_info.py
@@ -7,7 +7,7 @@ import boto3
 from prance import ResolvingParser
 from ruamel.yaml import YAML  # pylint: disable=import-error # type: ignore
 
-from ...services.import_agent.utils import clean_variable_name, fix_field
+from ....services.import_agent.utils import clean_variable_name, fix_field
 
 
 def get_clients(credentials, region_name="us-west-2"):
